# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.AddIncomingShareRequest:
    properties:
      inviteUrl:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.FileRow:
    properties:
      collectionId:
        type: string
      createdAt:
        type: string
      encryptedFileKey:
        type: string
      encryptedMetadata:
        type: string
      encryptedSizeBytes:
        type: integer
      fileKeyNonce:
        type: string
      id:
        type: string
      metadataNonce:
        type: string
      updatedAt:
        type: string
      uploaderUserId:
        type: string
    type: object
  handlers.IncomingShare:
    properties:
      canDelete:
        type: boolean
      canUpload:
        type: boolean
      createdAt:
        type: string
      encryptedCollectionKey:
        type: string
      encryptedName:
        type: string
      id:
        type: string
      nameNonce:
        type: string
      remoteServer:
        type: string
      uploadQuotaBytes:
        type: integer
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Self-hosted, end-to-end encrypted file storage with federation. All
    file content and metadata are encrypted client-side; the server stores only ciphertext.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kutup API
  version: 1.0.0
paths:
...
  /fed-proxy/{shareId}/files:
    get:
      parameters:
      - description: Incoming share UUID
        in: path
        name: shareId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.FileRow'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List files in a federated share (proxied)
      tags:
      - Federation Proxy
  /fed-proxy/{shareId}/files/{fileId}:
    delete:
      parameters:
      - description: Incoming share UUID
        in: path
        name: shareId
        required: true
        type: string
      - description: File UUID
        in: path
        name: fileId
        required: true
        type: string
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a file in a federated share (proxied)
      tags:
      - Federation Proxy
  /fed-proxy/{shareId}/files/{fileId}/download:
    get:
      parameters:
      - description: Incoming share UUID
        in: path
        name: shareId
        required: true
        type: string
      - description: File UUID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download a file from a federated share (proxied)
      tags:
      - Federation Proxy
  /fed-proxy/{shareId}/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Incoming share UUID
        in: path
        name: shareId
        required: true
        type: string
      - description: Encrypted metadata (base64)
        in: formData
        name: encryptedMetadata
        required: true
        type: string
      - description: Metadata nonce (base64)
        in: formData
        name: metadataNonce
        required: true
        type: string
      - description: Encrypted file key (base64)
        in: formData
        name: encryptedFileKey
        required: true
        type: string
      - description: File key nonce (base64)
        in: formData
        name: fileKeyNonce
        required: true
        type: string
      - description: Encrypted file content
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a file to a federated share (proxied)
      tags:
      - Federation Proxy
  /fed-proxy/incoming:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.IncomingShare'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List incoming federated shares
      tags:
      - Federation Proxy
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AddIncomingShareRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.IncomingShare'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept a federated share invite
      tags:
      - Federation Proxy
  /fed-proxy/incoming/{shareId}:
    delete:
      parameters:
      - description: Share UUID
        in: path
        name: shareId
        required: true
        type: string
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove an incoming federated share
      tags:
      - Federation Proxy
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"